Fix_6.1

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -4,6 +4,6 @@
 from typing import List, Union
 
 
-def sum_mixed_list(input_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(input_list: List[Union[int, float]]) -> float:
     """Sum mixed list"""
     return sum(input_list)
